+++ packages/components/src/styles/css-variables/spacing/static/index.ts
@@ -1,4 +1,5 @@
 export { spacingStatic2Xl } from './spacingStatic2Xl';
+export { spacingStatic2Xs } from './spacingStatic2Xs';
 export { spacingStaticLg } from './spacingStaticLg';
 export { spacingStaticMd } from './spacingStaticMd';
 export { spacingStaticSm } from './spacingStaticSm';