+++ packages/styles/projects/scss/src/spacing.ts
@@ -6,6 +6,7 @@ import {
   spacingFluidXl,
   spacingFluidXs,
   spacingStatic2Xl,
+  spacingStatic2Xs,
   spacingStaticLg,
   spacingStaticMd,
   spacingStaticSm,
@@ -15,6 +16,7 @@ import {
 
 export const getSpacingScss = () => {
   return `
+    $spacing-static-2xs: ${spacingStatic2Xs};
     $spacing-static-xs: ${spacingStaticXs};
     $spacing-static-sm: ${spacingStaticSm};
     $spacing-static-md: ${spacingStaticMd};